+++ lib/web_ui/lib/src/engine/safe_browser_api.dart
@@ -312,7 +312,7 @@ extension DecodeResultExtension on DecodeResult {
 @staticInterop
 class DecodeOptions {
   external factory DecodeOptions({
-    required int frameIndex,
+    required double frameIndex,
   });
 }
